requirements.txt
discord.py==2.2.3
requests==2.31.0
yt-dlp==2023.3.4
configparser==5.3.0
bs4